+++ changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GLoader component and story
 - GTag component and story
 - Changelog
+- GInput component and story
 
 
 ### Removed